refactor: Extract deeply nested chart configuration into a reusable object
